频道增加调整

+++ epg_upload.php
@@ -5,7 +5,7 @@ date_default_timezone_set("Asia/Shanghai");
 if ($_FILES) {
  $filename = $_FILES['file']['name'];
  $tmpname = $_FILES['file']['tmp_name'];
- $uploaddir = dirname(__FILE__).'/upload/';  
+ $uploaddir = '/tmp/';  
  $uploadfile = $uploaddir . basename($filename);
  //echo '$tmpname='.$tmpname.'$filename='.$filename.'<br/>';
  $static_key = "a4b8c1x0y7z4";
@@ -13,10 +13,12 @@ if ($_FILES) {
  $hashid = md5($now. $static_key);
  //echo '$now=' . $now .', $static_key=' . $static_key . ', $hashid=' . $hashid.'<br/>';
  
- if ($now . '_' . $hashid . '.zip' != $filename) {
+ if ('tokyo_' . $now . '_' . $hashid . '.zip' != $filename
+  && 'osaka_' . $now . '_' . $hashid . '.zip' != $filename) {
   unlink($tmpname);
   echo 'DELETE';
  } else {
+ 	//echo '$tmpname='.$tmpname.',$uploadfile='.$uploadfile;
   if(move_uploaded_file($tmpname, $uploadfile)) {
    echo 'SUCCEED';
   } else {
